remove example packages
These were sorta used for testing before the inttest package was created
and for better documentation this work a bit better as real Go examples
as defined by the testing package.
