--- examples/tls/certs/ca.conf
@@ -1,42 +1,0 @@
-[ ca ]
-default_ca = myca
-
-[ crl_ext ]
-# issuerAltName=issuer:copy  #this would copy the issuer name to altname
-authorityKeyIdentifier=keyid:always
-
- [ myca ]
- dir = ./
- new_certs_dir = $dir
- unique_subject = no
- certificate = $dir/ca.crt
- database = $dir/certindex
- private_key = $dir/ca.key
- serial = $dir/certserial
- default_days = 730
- default_md = sha1
- policy = myca_policy
- x509_extensions = myca_extensions
- crlnumber = $dir/crlnumber
- default_crl_days = 730
-
- [ myca_policy ]
- commonName = supplied
- stateOrProvinceName = supplied
- countryName = optional
- emailAddress = optional
- organizationName = supplied
- organizationalUnitName = optional
-
- [ myca_extensions ]
- basicConstraints = CA:false
- subjectKeyIdentifier = hash
- authorityKeyIdentifier = keyid:always
- keyUsage = digitalSignature,keyEncipherment
- extendedKeyUsage = serverAuth
- subjectAltName  = @alt_names
-
- [alt_names]
- DNS.1 = foobar.example 
- DNS.2 = *.foobar.example
-
