--- examples/tls/certs/ca.crt
@@ -1,23 +1,0 @@
------BEGIN CERTIFICATE-----
-MIID5TCCAs2gAwIBAgIUO3bMORsUSCu1ci96YZKjPp082u0wDQYJKoZIhvcNAQEL
-BQAwgYExCzAJBgNVBAYTAlVTMQswCQYDVQQIDAJDQTESMBAGA1UEBwwJU3Vubnl2
-YWxlMRQwEgYDVQQKDAtGb29iYXIgTHRkLjEXMBUGA1UEAwwOZm9vYmFyLmV4YW1w
-bGUxIjAgBgkqhkiG9w0BCQEWE3Jvb3RAZm9vYmFyLmV4YW1wbGUwHhcNMjIwNzI4
-MjE1MzAzWhcNNDkxMjEyMjE1MzAzWjCBgTELMAkGA1UEBhMCVVMxCzAJBgNVBAgM
-AkNBMRIwEAYDVQQHDAlTdW5ueXZhbGUxFDASBgNVBAoMC0Zvb2JhciBMdGQuMRcw
-FQYDVQQDDA5mb29iYXIuZXhhbXBsZTEiMCAGCSqGSIb3DQEJARYTcm9vdEBmb29i
-YXIuZXhhbXBsZTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAJgpghH9
-QhdqusW20suJv7VuwIAZS4Y+O+caKXYDu8YJO3dB0NtpKyLTN2tXAqv05qbCiDHh
-K4x5AtTA2fktFUMkX5x4KqG2Dq7l1G0NMlNao/EruxE54HxxEzcmSrj/d/b7Cfg1
-w20peqbM3eZxZel6hEmdqOymV04vlKYY/5YEug/idQs4uVkyvAXWaSADlRwMF47b
-TOTbQa6nY5Xvu+Ad3dYajCyiqjHHeBpqyOWZrfU3etzUgmG87umEz5duNU6w0H7w
-hNlfSDISIMiAL68BGXhF62/w3hbq/j9oBxEEncK+47C9eHcadYWc4PzbTzkVDfjq
-H3IGCOiLoa+pZUcCAwEAAaNTMFEwHQYDVR0OBBYEFB7pTgNG2PDFTHV7+fFveSCC
-XLEdMB8GA1UdIwQYMBaAFB7pTgNG2PDFTHV7+fFveSCCXLEdMA8GA1UdEwEB/wQF
-MAMBAf8wDQYJKoZIhvcNAQELBQADggEBAAeSSuFRjRuCJ+R+tqa6Au+TpVTywtk2
-qhQFilhPMU+uniYF1lTGxbpRQG9G+78h602zp3138AOJNxpTMj+5Tspo8TkyfUqj
-VIlv+LLYLG9MGqa+QUN52GERa+SNPVQ+ChbFhHr77eOodganv95DU+ytg95gpLp3
-iGbYSqYt1Eo8O4rFWBSO0YiH6tR4JcHk65QfKsTuS1hvVacAXTHMG1C57rqUG0JU
-ny16DQEBEbmUFKaZw/JOi1X5ZYfJ2kMs2RfTsZn9iAGF43J6BW6l0dds4bImuCYp
-aFCWbTGaZolSx9bH0LCVNtuHPFeb9eE+IvGgbr1AszLbeUjUpUxI+Oc=
------END CERTIFICATE-----
