--- examples/tls/certs/ca.key
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQCYKYIR/UIXarrF
-ttLLib+1bsCAGUuGPjvnGil2A7vGCTt3QdDbaSsi0zdrVwKr9Oamwogx4SuMeQLU
-wNn5LRVDJF+ceCqhtg6u5dRtDTJTWqPxK7sROeB8cRM3Jkq4/3f2+wn4NcNtKXqm
-zN3mcWXpeoRJnajspldOL5SmGP+WBLoP4nULOLlZMrwF1mkgA5UcDBeO20zk20Gu
-p2OV77vgHd3WGowsoqoxx3gaasjlma31N3rc1IJhvO7phM+XbjVOsNB+8ITZX0gy
-EiDIgC+vARl4Retv8N4W6v4/aAcRBJ3CvuOwvXh3GnWFnOD82085FQ346h9yBgjo
-i6GvqWVHAgMBAAECggEAK2ajgAgRwuYhceWMyJt4wAJgC+mrHhZANdHDg7La60iU
-2pY+QWcE88D1LLu7ulp3M5eHC1sjdcTKZu6+/ZC7YNROJS4tUsS/+Wj6p0Q9V4Mi
-AdcsW2R7wGLsTS2x4ISVri+6jz6eiFNtdYg3vtQXcnoMo4bBuQZVHqr5pkq5FAvC
-HRHgSkpFLaRT7vdT3UOqIwcIlZI2me5CGxrU1UgpwasG6O4pVjgg4+A4z82fgjko
-1nSUZhbH/XtXwWcuwX8r5T61Cq2i4mT+vEp6a9x/6yOM8YuRrWbG/7kGA2dscdZD
-0GlwoX6Lvm5vzrYoD4v60nRAlZacS5i8S091Z/6hcQKBgQDWsejDpaoIHyFx/ig8
-AOIkDbP4XWdZ8gEkkb4I88WeV/a8GBfmCuwTHPw4uyfuEVY+qvt4x3f3PliE4AV5
-0B+hG6d5i/Xa336dafHTppsSwUMxAdmnG0GEf+0Eg50l9JlGSWtWxL5RBFZgbvOx
-VADkjC7E3zFtX7dIxM5A5/aALwKBgQC1b8Bm6FKsrmm8saUgTwdD2mkb4miJ64Nw
-EnMOygfWGelOCEb0rkkjdF6cEq/HwRFPcNW1b/ACg3+7f4yB6HomoklACDz3oElr
-CpzAW3E6rdkv8vG4p/+4YSgo3aP113SXL/V+LFxRU+wjCqoiXadQ07lIiom1sqXa
-HkrkWufOaQKBgE3ohZTpuwfzdAmPtz3+/BElP9ey3VxQa4W/tXw+KcpKFWkwPQvn
-jr9o10Ym9Fh5xhg7p8GucKQx5aFiQdpwYenuxRlgZrkkPSgWWAD8Bm4J374a88am
-J2JTocjvkclCN5S/88mq330627Vmzdvdpe1JyXb7n04XW8V2VXpMMrFVAoGAD/8E
-RY5iC5ts8xfuaQTOrSrKbYzujopXEkBuKHWjbzrlmN2d8+CRzjBqTvROzxu0fP6D
-uPVfgbk6P7ulFab7bgCM2LphNavKDUfoPgVR3EyLg3C14wPcxdK7Ca9Qz1Hpk5LE
-4BTlJr5xaT19dn4vTC7+fOjJ710Kxko5bByfp1ECgYEAofZjYtJGq9gZOYkL/5cJ
-4HuVxkkAi+H8OvRWIdY9Cq0vcXM4eyYc2bW7qA0G6X6+dlq4S8sTJO6GHGxF/alo
-RpzKr/3ToSUgJU5Ufe4WRyp2TDEI0PpFVBnk/K0SdBShmZY3AuXM2OovhQnJbyx5
-ocWkOUDs8wlZ+P6rA/0ZkAk=
------END PRIVATE KEY-----
