--- examples/tls/certs/certserial
@@ -1,1 +1,0 @@
-01
