--- examples/tls/certs/client.crt
@@ -1,22 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDlzCCAn8CFEWBCuc/93ezY29cddUdP5B2W4zBMA0GCSqGSIb3DQEBCwUAMIGB
-MQswCQYDVQQGEwJVUzELMAkGA1UECAwCQ0ExEjAQBgNVBAcMCVN1bm55dmFsZTEU
-MBIGA1UECgwLRm9vYmFyIEx0ZC4xFzAVBgNVBAMMDmZvb2Jhci5leGFtcGxlMSIw
-IAYJKoZIhvcNAQkBFhNyb290QGZvb2Jhci5leGFtcGxlMB4XDTIyMDcyODIxNTQ0
-NloXDTQ5MTIxMjIxNTQ0NlowgY0xCzAJBgNVBAYTAlVTMQswCQYDVQQIDAJDQTES
-MBAGA1UEBwwJU3Vubnl2YWxlMRQwEgYDVQQKDAtGb29iYXIgTHRkLjEiMCAGA1UE
-AwwZZ28tbmV0Y29uZi5mb29iYXIuZXhhbXBsZTEjMCEGCSqGSIb3DQEJARYUcm9v
-dEBmb290YmFyLmV4YW1wbGUwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIB
-AQDCB4rbBzX0gjBRs1x7oEFCIhwmxqrSz+kAP8jwU+lRj88mv9z/AiHYIj96Z2/Z
-97c7keBRQ6+8IupmmFwHkoo10PV+IogMRxWRqePxtxTjSIrkeX859clCkpQDx8Wc
-h4wIvfoT7QusnhpPCCa64ZMpdj2g0CyjHtdiVBALTymqWysDvr51KMf/AaTAHf25
-McZyiZcXjHv0MoR8f/ZN0vBv/L2xsf5cMZHdwlPbzBq7kLLFr38bfvcOrmD5f5hZ
-kkrXuu2PBuy8bhW791FXh8zEeIKgDNQhfok7ciTOmNN8+KP40y5qmEEZ3TYzeEfH
-fSIrcbEXIeg3wJk9mOMqwECTAgMBAAEwDQYJKoZIhvcNAQELBQADggEBACH0S00T
-EJ2xtrfvbhv3c/O22DPGqrzAoF93P/hi0yWvZjdLOnJYR3HBdH9jdhlH99dJWJlY
-Qf9txbOYA410RGzTaugXRa/i2NhixaQAXvRPCO69GMzYWaXrUe4YCzrG4b+7ZM9L
-ZQxnLwUpQwNxHLLL53nLsikim5sxaX3z9tqKpVNqBo52Lwq5dB6BojCT3fmvvIZY
-zNWqtB76EV5c0mo+LTj4fcFLO15ACUkxiJfcc0r3DcRbNY915muIGIkjQpFp65U+
-Xp9cNYP33kRG67TFJm+/4CzB2ChZSHAkuN7LuKhu9ZWFSy13JzBsBSm1hiM5mZ2y
-aCtePNC0645i43U=
------END CERTIFICATE-----
