--- examples/tls/certs/client.csr
@@ -1,18 +1,0 @@
------BEGIN CERTIFICATE REQUEST-----
-MIIC0zCCAbsCAQAwgY0xCzAJBgNVBAYTAlVTMQswCQYDVQQIDAJDQTESMBAGA1UE
-BwwJU3Vubnl2YWxlMRQwEgYDVQQKDAtGb29iYXIgTHRkLjEiMCAGA1UEAwwZZ28t
-bmV0Y29uZi5mb29iYXIuZXhhbXBsZTEjMCEGCSqGSIb3DQEJARYUcm9vdEBmb290
-YmFyLmV4YW1wbGUwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDCB4rb
-BzX0gjBRs1x7oEFCIhwmxqrSz+kAP8jwU+lRj88mv9z/AiHYIj96Z2/Z97c7keBR
-Q6+8IupmmFwHkoo10PV+IogMRxWRqePxtxTjSIrkeX859clCkpQDx8Wch4wIvfoT
-7QusnhpPCCa64ZMpdj2g0CyjHtdiVBALTymqWysDvr51KMf/AaTAHf25McZyiZcX
-jHv0MoR8f/ZN0vBv/L2xsf5cMZHdwlPbzBq7kLLFr38bfvcOrmD5f5hZkkrXuu2P
-Buy8bhW791FXh8zEeIKgDNQhfok7ciTOmNN8+KP40y5qmEEZ3TYzeEfHfSIrcbEX
-Ieg3wJk9mOMqwECTAgMBAAGgADANBgkqhkiG9w0BAQsFAAOCAQEAgQ1jAzz0Pi3S
-hF+KKYxUZduCbA1hBT6YVub38cThocP4MOvymS7pwFgBR5jvfaynxBxVKNy1brmR
-aj8bu7VC+NBZ59e781RZjKzx4kk697uwO+txk3QaGFWTMFb6G3H2CFKhM/Bqyf/L
-mdHlxqkMjfX9UtMl83ZN7Yy6Kx9nW2GgPveTXsjLVTISgXfd+L7Mv/JIsgT8ZlNU
-GgbyinnkfnBnuwioO0j9d2FoYuMchN98vXEsccPMJsaiU8U69lqgSwqnPS0MLTKL
-ZN4Rq2A4PzunW2LGMESO1Z7kvE9q9gBKfCXgfSLo4R9d6nqJ/6LzWzi2mB/IqdYG
-TGMfrJbylw==
------END CERTIFICATE REQUEST-----
