--- examples/tls/certs/client.key
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDCB4rbBzX0gjBR
-s1x7oEFCIhwmxqrSz+kAP8jwU+lRj88mv9z/AiHYIj96Z2/Z97c7keBRQ6+8Iupm
-mFwHkoo10PV+IogMRxWRqePxtxTjSIrkeX859clCkpQDx8Wch4wIvfoT7QusnhpP
-CCa64ZMpdj2g0CyjHtdiVBALTymqWysDvr51KMf/AaTAHf25McZyiZcXjHv0MoR8
-f/ZN0vBv/L2xsf5cMZHdwlPbzBq7kLLFr38bfvcOrmD5f5hZkkrXuu2PBuy8bhW7
-91FXh8zEeIKgDNQhfok7ciTOmNN8+KP40y5qmEEZ3TYzeEfHfSIrcbEXIeg3wJk9
-mOMqwECTAgMBAAECggEABpwsgNW4DYJklFlLQfLtsOEUmxroNZJbr7nR5fjAhBol
-cSksJYvfDKFDZFojF1vhuX0jaov8b/5d2ezq2wnaUjGyNf/bKVU/84cZeuHv1ju3
-GkgsAL981xsYlnW3J7lTw0slj97i0kNN8M9S8nQ+Moq9WBMqJ3ck2KFggozJ20HR
-/9GxnMfOLvB4p0qiHZkGc4q7eLFBdZb3NWsHM0Ag5l9D4vlXxDKeLL/M6OquXTcA
-PMSv5jvgeDIhQalwyxERp9WpaLSUsgPJCYBfwubWA85kSzHdJOloBQaIN75VDBxq
-FmeKzQF796FKG3ckx9GAdlS57bfT+dq5tsnC6FaWcQKBgQDfwFuFI5XegqCDeZcc
-7yCSEgMOuRDA4Lp8z5L8jCQ2Zb0tLbDwiFtr2P6wOME0cPMOf9c9Pg9w0mrMaAyv
-vI0f4ztEC0tJEoa70KFNlrn6lFd0oAiNj4vO2c9Yxiq8Yn1MXpjSfZAOee9pFkKt
-Pz9wmCoJnwamlqBXCFGwKKomnQKBgQDd/oy4TwqpHSK5fwWbvKK9+1XM2XgydLqr
-04BITkHvL3S80uch1MlDhVch6UMW63GFkfM9oa9Fjud0o8qIstJUDZK0ySRui6eO
-NcfEPMx+LrThLUBlSWtSODfd43I0LTHvs7mlyACEv0dEWf09Fnb99XJHQXqwgT8J
-i3yqjM7E7wKBgC5agltFsQGk3C4AnY6v93ZrZtjA/DiJFflPY2sQeVwPna5jXdmm
-1hG0pabKSzRbBe+GCog0eCxuWxO3ppEtparqf1esY09jGCiGozQYurSv0KJnmP4C
-s/sJYoxImYfne3+7xTsuKg7lsfR6WRvdE35295aArHcMk3EUVyAAfKd9AoGAJFpG
-2Pe4mzD2OsIePP6K9+lsh3slP731w6WC4FRQX8a5qyIuT0/1FWWEez0ji2FK2cCo
-wHL1rMJWDLe1tuTxpGoEmC5IFQIxkgwkhnIeWptza1xpqyx2sqQaIe/ZB8pO3x3j
-TM6Ppy1bkknaOmKvYa8L47dJTBYeKIX7CEK+qWkCgYEAtILZ43i1N1hFbp6WGES+
-+b/QvIa7b47P1FY1JBhCX61BX/Ihyw5/mnQhnx+b8dABKvi7onxcToT3KEmSEhHr
-3VIjDg6qY1XlsG1zVPtZt/srruuE0SjmRPXG29llKIcMPp7Yv1GT4xRE4Vt6cvw8
-q45WC1cb8myRFFrvbuxWVPw=
------END PRIVATE KEY-----
