--- examples/tls/certs/crlnumber
@@ -1,1 +1,0 @@
-02
