--- examples/tls/certs/crlnumber.old
@@ -1,1 +1,0 @@
-01
