--- examples/tls/certs/device.crt
@@ -1,22 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDkzCCAnsCFHE+NqEGg/awYEJVCPJfma6MAxa4MA0GCSqGSIb3DQEBCwUAMIGB
-MQswCQYDVQQGEwJVUzELMAkGA1UECAwCQ0ExEjAQBgNVBAcMCVN1bm55dmFsZTEU
-MBIGA1UECgwLRm9vYmFyIEx0ZC4xFzAVBgNVBAMMDmZvb2Jhci5leGFtcGxlMSIw
-IAYJKoZIhvcNAQkBFhNyb290QGZvb2Jhci5leGFtcGxlMB4XDTIyMDcyODIxNTMx
-MFoXDTQ5MTIxMjIxNTMxMFowgYkxCzAJBgNVBAYTAlVTMQswCQYDVQQIDAJDQTES
-MBAGA1UEBwwJU3Vubnl2YWxlMRQwEgYDVQQKDAtGb29iYXIgTHRkLjEfMB0GA1UE
-AwwWcm91dGVyMS5mb29iYXIuZXhhbXBsZTEiMCAGCSqGSIb3DQEJARYTcm9vdEBm
-b29iYXIuZXhhbXBsZTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBANWI
-2X3jKrI+mjlcRHo9tCv+A1Z3oyKXnnTEDgkcVCKwxjlalLAj3iCt0Dx9595C/+Zs
-y2PlGAyirLVAwKnqOGY7rdlIsBKsgfock0zmAZmv/1LFswP9Je6Bgpk7jlJ/OWJJ
-TCL2pJsrH1sNYEIb8OrjM9ooEKFTghTaSp18LdL/iH+2GpPWMtKmSdqYjHOcTOeM
-P/vsCbZLZR8FmKxdC17PbJs3iMCbv4eJQxedPtaFySKaTyyc7Eo/CY0LmvT4VnjY
-HrYJvZnDCVd1E414qVV4uG416oEzA/ZNeFOI4uVl/buq4R/3rPRxENgEoZNELTcb
-hQbR1mYyMqrLESECIE0CAwEAATANBgkqhkiG9w0BAQsFAAOCAQEACYCoL/rNVcHI
-Hl3DOvkL4EcjZf9mGjXkHq3dMlylDJW/veePQeTEfDS9DbGGjy7wVollOSVEQujH
-MhguKd31UmlWP1+8El8ePIF6/dW7MuX9sXMuzz6/xyIH+WckAEZoZ7W5t5lLYz7N
-YTAPn2E14QENaLELWyPJsb5lBhHUFp6kuTOErixP97w+x48+/ytfeX5QEN5H2+hY
-X1B8VZu1aQri9m7l5Ak819P6e75xgzc5qbJ3GofIdiZxLD1s8yPXoOgB6aMmtNSG
-WuRf5NAX2vpV4eCgjc5WvdXwLm1YjkfcF4Tkm2nl0FAK5QObYYq6e4FZVWVW9esu
-gLsMUZT6/w==
------END CERTIFICATE-----
