--- examples/tls/certs/device.csr
@@ -1,18 +1,0 @@
------BEGIN CERTIFICATE REQUEST-----
-MIICzzCCAbcCAQAwgYkxCzAJBgNVBAYTAlVTMQswCQYDVQQIDAJDQTESMBAGA1UE
-BwwJU3Vubnl2YWxlMRQwEgYDVQQKDAtGb29iYXIgTHRkLjEfMB0GA1UEAwwWcm91
-dGVyMS5mb29iYXIuZXhhbXBsZTEiMCAGCSqGSIb3DQEJARYTcm9vdEBmb29iYXIu
-ZXhhbXBsZTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBANWI2X3jKrI+
-mjlcRHo9tCv+A1Z3oyKXnnTEDgkcVCKwxjlalLAj3iCt0Dx9595C/+Zsy2PlGAyi
-rLVAwKnqOGY7rdlIsBKsgfock0zmAZmv/1LFswP9Je6Bgpk7jlJ/OWJJTCL2pJsr
-H1sNYEIb8OrjM9ooEKFTghTaSp18LdL/iH+2GpPWMtKmSdqYjHOcTOeMP/vsCbZL
-ZR8FmKxdC17PbJs3iMCbv4eJQxedPtaFySKaTyyc7Eo/CY0LmvT4VnjYHrYJvZnD
-CVd1E414qVV4uG416oEzA/ZNeFOI4uVl/buq4R/3rPRxENgEoZNELTcbhQbR1mYy
-MqrLESECIE0CAwEAAaAAMA0GCSqGSIb3DQEBCwUAA4IBAQDPQm2Sav3/mN28sPAF
-C1CBgPCKIlpQHhGiNWIkAZDmPNzg/DS8LZol/7Iu8YhGRmbwgO3pGEYf/n+44lsa
-gbcMBkg8rvCq5guoAIb18faHzai+OkRb5bYO3qer1eFMoskBd9SEApcL+srVbbcw
-nzecKdhAKbsvv+vYz2tXQ/k0filkf0iZFehqx7bnxlUidfdFvoetMwCo6vCBcmJ8
-hGXAgRSABHz0ZifA24lXcLJHjWbzXosRocC/7mtb1nIHb6bIkZ3rAA+Bcc3ObXrK
-VY8gIvWXWKPYNmCSpGVdHsGU3bOvLlAPE0VmdB1fAyWbqNxtwjvnanV8Lwja2Qw4
-RKaV
------END CERTIFICATE REQUEST-----
