--- examples/tls/certs/device.key
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDViNl94yqyPpo5
-XER6PbQr/gNWd6Mil550xA4JHFQisMY5WpSwI94grdA8fefeQv/mbMtj5RgMoqy1
-QMCp6jhmO63ZSLASrIH6HJNM5gGZr/9SxbMD/SXugYKZO45SfzliSUwi9qSbKx9b
-DWBCG/Dq4zPaKBChU4IU2kqdfC3S/4h/thqT1jLSpknamIxznEznjD/77Am2S2Uf
-BZisXQtez2ybN4jAm7+HiUMXnT7Whckimk8snOxKPwmNC5r0+FZ42B62Cb2ZwwlX
-dRONeKlVeLhuNeqBMwP2TXhTiOLlZf27quEf96z0cRDYBKGTRC03G4UG0dZmMjKq
-yxEhAiBNAgMBAAECggEARVrLeU3dQkN/B/5l52Y/LMK2BtGY1NV2C+rZvC/NS+8q
-ktpOzN75eJe7uMoUE2ajgw33CZ0RMJltivlbGlUOuHXNw3MObSJGnRvG5AUYo/g7
-74By4mrIdSzvZCGp2jWkR74S0pBVamiZKJOxngdBOqlM4CB/NkbPS0KTVY9NuufY
-UQzVROKgGRgEHjGshl2bN326te14gPlCQGgXYy0BScXCV1eW3AKSYY7FekiSLmLp
-btJZoapqHeDiGK8MnUQl5DmeiLuK3c4ewEH2Twc2PNzTiUFm/AuOCNBaQoZ6vqeQ
-Nj2+vSgLtAzqoZQQRN56yrod82TGTxG8XxUx4smaqQKBgQDxv9aICm28vybwk/ua
-7n2YoaayxDwN1UX3JK3THLiyO44EVHUcrCNkanP1qI7qxK13kHD6OG0AWvi88/ap
-93QXum9zQ5iR+6P489RuGZryaaweqVc48hPKLIJUfKQmI03HX1GwLy39HxzX5gab
-g7Wv30P5Sw0vgEkSuYlIiHJKrwKBgQDiHzsp7poSTmD0auNWyMGlkC+oWpYxiWXL
-9VwUarcg1+6Kx507tqURAO1KDbxG2lc59ChQ788V2y2SmTZcJO5OeF8QM85hyy98
-XDXkAOWQlEGXaSrLqeMLEnpnazjDuFNAwQ90sCWh7edx5dAbjaKmggtx4cmuw/Je
-NOyffSvTwwKBgQDp1kYNEQlUqRFhY8Lb+hRCrxovkOEGAGXMyQ3aRs3gmPql9YvS
-UC9r0gyYcpLLQ7MHFBAz+KB214gCPXk6rsD8srlbJOdPTbkt9axlUx8tnEmA3wwf
-aQj7ZlIqpuWgfkWds/WwFfLGBHJB455AjwYPOEbnvFDkwZHtKso//ruOwQKBgAcY
-5dl+FLXoYeCqDjKmIHXa3bQSC87IOjq5m8dERjUwsdsHQuch53JuLgR6A+PRbJV0
-FWPc8woDlB2a03r26cfqvUkb5ylooDfuCP78DDiKfriK4336+hZZ4R+n3OM/wNUy
-ehK2iKfZeL85cFu9CVLimAteA9+JZO0SCi9DVYHTAoGAaWzJy9KXk9LjawadbFNQ
-++KDD7v3oV5sG2VWe/ju0jLNb4JiBMg2TFcQ+o1SdKKqg4doPRuG6SaWhoFI80EH
-o3bIApWMzhz+RDpyjg/gEUiCy5cChV1Jo7Uj/8M5jO5duB/9zW3bjiu/24j0ojxz
-CfwRhjNrasAaR1tQ2tK7lO4=
------END PRIVATE KEY-----
